Adds a unit test and fixes an initializer.
