+++ AztecTests/NSAttributedString/Conversions/AttributedStringParserTests.swift
@@ -996,25 +996,6 @@ class AttributedStringParserTests: XCTestCase {
         
         XCTAssertEqual(textNode.text(), captionText)
     }
-    
-    /// Verifies that Images with *Styled* caption fields get properly converted into their corresponding HTML
-    ///
-    /// - Input: Image Attachment, with a caption in Bold + Italics.
-    ///
-    /// - Output: <figure><img src="."><figcaption><strong><em>Bold and Italics</em></strong></figcaption></figure>
-    ///
-    func testAppendProperty() {
-        let string = NSMutableAttributedString(string: "Hello world!", attributes: Constants.sampleAttributes)
-        
-        string.append(paragraphProperty: Figure())
-        
-        guard let paragraphStyle = string.attribute(.paragraphStyle, at: 0, effectiveRange: nil) as? ParagraphStyle else {
-            XCTFail() // Expected a ParagraphStyle object.
-            return
-        }
-        
-        XCTAssert(paragraphStyle.hasProperty(where: { $0 is Figure }))
-    }
 }
 
 // MARK: - Helpers
